break up the Python API into components
